add district generated json file
