--- frontend/lib/models/district_model.dart
@@ -1,10 +1,0 @@
-// use JsonSerializable
-class District {
-  final String id;
-  final String name;
-
-  District({
-    required this.id,
-    required this.name,
-  });
-}
