+++ frontend/lib/pages/admin/admin_page.dart
@@ -1,7 +1,7 @@
 import 'dart:math' as math;
 
 import 'package:flutter/material.dart';
-import 'package:orot/models/user_model.dart';
+import 'package:orot/models/user.dart';
 import 'package:orot/pages/admin/add_coordinator_page.dart';
 import 'package:orot/pages/admin/add_district_page.dart';
 import 'package:orot/pages/admin/add_family_page.dart';
