+++ frontend/lib/services/admin_service.dart
@@ -1,7 +1,7 @@
 import 'package:cloud_functions/cloud_functions.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-import 'package:orot/models/district_model.dart';
+import 'package:orot/models/district.dart';
 
 class AdminService {
   final FirebaseFunctions _functions = FirebaseFunctions.instance;
