+++ frontend/pubspec.yaml
@@ -50,6 +50,7 @@ dependencies:
   flutter_native_splash: ^2.4.3
   scaffold_gradient_background: ^1.0.3+1
   provider: ^6.1.2
+  json_serializable: ^6.9.2
 
 dev_dependencies:
   flutter_test:
@@ -62,7 +63,7 @@ dev_dependencies:
   # rules and activating additional ones.
   flutter_lints: ^5.0.0
   build_runner: ^2.4.14
-
+  
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
